Trait frc42_dispatch::hash::Hasher

pub trait Hasher {
    // Required method
    fn hash(&self, bytes: &[u8]) -> Vec<u8> ;
}

Minimal interface for a hashing function

Hasher::hash() must return a digest that is at least 4 bytes long so that it can be cast to a u32

Required Methods

fn hash(&self, bytes: &[u8]) -> Vec<u8>

For an input of bytes return a digest that is at least 4 bytes long